backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
django-redis==5.4.0
redis==5.0.7
celery==5.4.0
python-dotenv==1.0.1
dj-database-url==2.2.0
cryptography==42.0.8
bcrypt==4.1.3
django-cors-headers==4.4.0
gunicorn==22.0.0
safety==3.2.4
bandit==1.7.9
